# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Local Doujin Studio",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/render": {
      "post": {
        "summary": "Render Project",
        "operationId": "render_project_api_projects__project_id__render_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Revision"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RenderRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMutationResponse_ProjectRenderResult_"
                }
              }
            }
          },
          "409": {
            "description": "revision競合、または生成キャンセル（入力変更・構成置換など）",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ProjectRevisionConflictResponse"
                    }
                  ],
                  "title": "Response 409 Render Project Api Projects  Project Id  Render Post"
                }
              }
            }
          },
          "502": {
            "description": "画像生成バックエンドが失敗した",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/panels/{panel_id}/generate-image": {
      "post": {
        "summary": "Generate Panel Image",
        "operationId": "generate_panel_image_api_projects__project_id__panels__panel_id__generate_image_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "panel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Panel Id"
            }
          },
          {
            "name": "revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Revision"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/GenerationJobCreate"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMutationResponse_PanelImageGenerationResult_"
                }
              }
            }
          },
          "409": {
            "description": "revision競合、または生成キャンセル（入力変更・構成置換など）",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ProjectRevisionConflictResponse"
                    }
                  ],
                  "title": "Response 409 Generate Panel Image Api Projects  Project Id  Panels  Panel Id  Generate Image Post"
                }
              }
            }
          },
          "502": {
            "description": "画像生成バックエンドが失敗した",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/panels/{panel_id}/prompt-preview": {
      "get": {
        "summary": "Preview Panel Prompt",
        "operationId": "preview_panel_prompt_api_projects__project_id__panels__panel_id__prompt_preview_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "panel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Panel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/panels/{panel_id}/generation-jobs": {
      "post": {
        "summary": "Create Generation Job",
        "operationId": "create_generation_job_api_projects__project_id__panels__panel_id__generation_jobs_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "panel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Panel Id"
            }
          },
          {
            "name": "revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Revision"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/GenerationJobCreate"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMutationResponse_GenerationJobResponse_"
                }
              }
            }
          },
          "409": {
            "description": "revision競合。最新のprojectを同梱する",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRevisionConflictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/generation-jobs": {
      "post": {
        "summary": "Create Batch Generation Jobs",
        "operationId": "create_batch_generation_jobs_api_projects__project_id__generation_jobs_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Revision"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/BatchGenerationJobCreate"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMutationResponse_BatchGenerationJobResult_"
                }
              }
            }
          },
          "409": {
            "description": "revision競合。最新のprojectを同梱する",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRevisionConflictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Generation Jobs",
        "operationId": "list_generation_jobs_api_projects__project_id__generation_jobs_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GenerationJobResponse"
                  },
                  "title": "Response List Generation Jobs Api Projects  Project Id  Generation Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/generation-jobs/{job_id}": {
      "get": {
        "summary": "Get Generation Job",
        "operationId": "get_generation_job_api_generation_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/generation-jobs/{job_id}/cancel": {
      "post": {
        "summary": "Cancel Generation Job",
        "operationId": "cancel_generation_job_api_generation_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMutationResponse_GenerationJobResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/panels/{panel_id}/candidates/{candidate_id}/select": {
      "post": {
        "summary": "Select Panel Candidate",
        "operationId": "select_panel_candidate_api_projects__project_id__panels__panel_id__candidates__candidate_id__select_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "panel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Panel Id"
            }
          },
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          },
          {
            "name": "revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMutationResponse_PanelPageRenderResult_"
                }
              }
            }
          },
          "409": {
            "description": "revision競合。最新のprojectを同梱する",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRevisionConflictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/panels/{panel_id}/use-stub": {
      "post": {
        "summary": "Use Stub Panel Image",
        "operationId": "use_stub_panel_image_api_projects__project_id__panels__panel_id__use_stub_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "panel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Panel Id"
            }
          },
          {
            "name": "revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMutationResponse_PanelImageGenerationResult_"
                }
              }
            }
          },
          "409": {
            "description": "revision競合。最新のprojectを同梱する",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRevisionConflictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ApiErrorResponse",
        "description": "FastAPI標準のエラー本体({\"detail\": ...})に対応する型。\n\n409(キャンセル)・502(生成バックエンド失敗)などの同期API契約をOpenAPIへ明示するため。"
      },
...
      "BatchGenerationJobCreate": {
        "properties": {
          "page": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "candidate_count": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Candidate Count",
            "default": 1
          }
        },
        "type": "object",
        "title": "BatchGenerationJobCreate"
      },
...
      "GenerationJobCreate": {
        "properties": {
          "candidate_count": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Candidate Count",
            "default": 1
          }
        },
        "type": "object",
        "title": "GenerationJobCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectMutationResponse_BatchGenerationJobResult_": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectDetail"
          },
          "result": {
            "$ref": "#/components/schemas/BatchGenerationJobResult"
          }
        },
        "type": "object",
        "required": [
          "project",
          "result"
        ],
        "title": "ProjectMutationResponse[BatchGenerationJobResult]"
      },
...
      "ProjectMutationResponse_GenerationJobResponse_": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectDetail"
          },
          "result": {
            "$ref": "#/components/schemas/GenerationJobResponse"
          }
        },
        "type": "object",
        "required": [
          "project",
          "result"
        ],
        "title": "ProjectMutationResponse[GenerationJobResponse]"
      },
...
      "ProjectMutationResponse_PanelImageGenerationResult_": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectDetail"
          },
          "result": {
            "$ref": "#/components/schemas/PanelImageGenerationResult"
          }
        },
        "type": "object",
        "required": [
          "project",
          "result"
        ],
        "title": "ProjectMutationResponse[PanelImageGenerationResult]"
      },
      "ProjectMutationResponse_PanelPageRenderResult_": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectDetail"
          },
          "result": {
            "$ref": "#/components/schemas/PanelPageRenderResult"
          }
        },
        "type": "object",
        "required": [
          "project",
          "result"
        ],
        "title": "ProjectMutationResponse[PanelPageRenderResult]"
      },
      "ProjectMutationResponse_ProjectRenderResult_": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectDetail"
          },
          "result": {
            "$ref": "#/components/schemas/ProjectRenderResult"
          }
        },
        "type": "object",
        "required": [
          "project",
          "result"
        ],
        "title": "ProjectMutationResponse[ProjectRenderResult]"
      },
...
      "ProjectRenderResult": {
        "properties": {
          "page_assets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Page Assets"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "page_assets"
        ],
        "title": "ProjectRenderResult"
      },
      "ProjectRevisionConflictResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "code": {
            "type": "string",
            "const": "project_revision_conflict",
            "title": "Code",
            "default": "project_revision_conflict"
          },
          "expected_revision": {
            "type": "integer",
            "title": "Expected Revision"
          },
          "actual_revision": {
            "type": "integer",
            "title": "Actual Revision"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectDetail"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "expected_revision",
          "actual_revision",
          "project"
        ],
        "title": "ProjectRevisionConflictResponse",
        "description": "revision競合時の専用409応答。最新projectを同梱する。"
      },
...
      "PromptPreviewResponse": {
        "properties": {
          "panel_id": {
            "type": "string",
            "title": "Panel Id"
          },
          "positive_prompt": {
            "type": "string",
            "title": "Positive Prompt"
          },
          "negative_prompt": {
            "type": "string",
            "title": "Negative Prompt"
          },
          "character_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Character Ids"
          }
        },
        "type": "object",
        "required": [
          "panel_id",
          "positive_prompt",
          "negative_prompt"
        ],
        "title": "PromptPreviewResponse"
      },
...
      "RenderRequest": {
        "properties": {
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "title": "RenderRequest"
      },
...
    }
  }
}